+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - PostgreSQL: Add force TLS/SSL and other logging flags
 - PostgreSQL: Add explicit log stream CloudWatch resource
 - Valkey: Upgrade default version from v8.1 to v8.2
+- Bastion: Add Debian 13 Trixie
 - Bastion: Upgrade Node.js to v24.x
 - Bastion: Upgrade Golang from v1.25.0 to v1.25.4
 - Bastion: Add RDS CA Certificate for PostgreSQL